chore: move get_lowest_stateless_rate to utils
