+++ tests/test_beta_rate.py
@@ -1,6 +1,7 @@
 import pytest
 
 import easypost
+from easypost.util import get_lowest_stateless_rate
 
 
 @pytest.mark.vcr()
@@ -16,6 +17,6 @@ def test_beta_get_lowest_stateless_rate(basic_shipment):
     """Tests that we can return the lowest stateless rate from a list of stateless rates."""
     stateless_rates = easypost.beta.Rate.retrieve_stateless_rates(**basic_shipment)
 
-    lowest_stateless_rate = easypost.beta.Rate.get_lowest_stateless_rate(stateless_rates)
+    lowest_stateless_rate = get_lowest_stateless_rate(stateless_rates)
 
     assert lowest_stateless_rate["service"] == "First"
